Moved this file from EducationStyles2006 product.  This is the same as the Plone default except it contains one additional line:
var promo_folder = '<tal:block replace="here/COEimages/absolute_url" />';

This causes an error if the COEimages folder does not exist.  By moving it to 
PromotionFlash/skins/PromotionFlash/plone_javascript_variables.js.pt,
this allows the EducationStylesProduct to be installed without the PromotionFlash project.  It is not used in the CNE site, which uses the EducationalStyles2006 product. 

JEB6