Revised the tool route root factory to deliver a traversable context.  This allows for further traversal after the tool route has been established.  This is a hybrid of url dispatch and traversal.
Moved configuration of the tools into individual configuration.zcml files, which is preparation for tool and framework separation.  

Revised the tools in response to the changes to the root factory.  Added additional views to the todo tool to see the effects of traversal on todo items.  

Removed the zope.dublincore from the todo model, because the zope.dublincore uses zope.security.  Until I figure out a way around that, zope.dublincore will be off limits.